feat(backend): token control for ticket

+++ backend/src/main/java/buzzspire/helpdesk/business/concreates/TicketManager.java
@@ -7,9 +7,14 @@ import buzzspire.helpdesk.dataAccess.abstracts.TicketDAO;
 import buzzspire.helpdesk.dto.request.ticket.TicketRequest;
 import buzzspire.helpdesk.entities.concreates.Ticket;
 import buzzspire.helpdesk.entities.concreates.TicketPriority;
+import buzzspire.helpdesk.entities.concreates.TicketStatus;
 import buzzspire.helpdesk.entities.concreates.User;
+import buzzspire.helpdesk.security.JwtTokenProvider;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+import java.util.Date;
 import java.util.List;
 
 @Service
@@ -17,24 +22,30 @@ public class TicketManager implements TicketServices {
 
     // this field injection
     private final TicketDAO ticketDAO;
+    private final JwtTokenProvider jwtTokenProvider;
 
     // constructor injection
-    public TicketManager(TicketDAO ticketDAO) {
+    public TicketManager(TicketDAO ticketDAO, JwtTokenProvider jwtTokenProvider) {
         this.ticketDAO = ticketDAO;
+        this.jwtTokenProvider = jwtTokenProvider;
     }
 
 
     // this method is used to add a ticket
     @Override
-    public Result addTicket(TicketRequest ticket) {
+    public Result addTicket(TicketRequest ticket, String token) {
+        if (!jwtTokenProvider.validateToken(token)){
+            return new Result(false, "Token is not valid");
+        }
+        long userId = jwtTokenProvider.getIdFromToken(token);
         try {
             this.ticketDAO.save(Ticket.builder()
                     .title(ticket.getTitle())
                     .description(ticket.getDescription())
                     .priority(TicketPriority.builder().id(ticket.getPriority()).build())
-                    .user(User.builder().id(ticket.getUser()).build())
-                    .status(ticket.getStatus())
-                    .date(ticket.getDate())
+                    .user(User.builder().id(userId).build())
+                    .status(TicketStatus.builder().id(ticket.getStatus()).build())
+                    .date(new Date())
                     .build());
         } catch (Exception e) {
             return new Result(false, "Ticket not added");
@@ -44,7 +55,14 @@ public class TicketManager implements TicketServices {
 
     // this method is used to delete a ticket
     @Override
-    public Result deleteTicket(long id) {
+    public Result deleteTicket(long id, String token) {
+        if (!jwtTokenProvider.validateToken(token)){
+            return new Result(false, "Token is not valid");
+        }
+        String role = jwtTokenProvider.getRoleFromToken(token);
+        if (!role.equals("ADMIN")){
+            return new Result(false, "You are not authorized");
+        }
         try {
             this.ticketDAO.deleteById(id);
         } catch (Exception e) {
@@ -73,7 +91,14 @@ public class TicketManager implements TicketServices {
 
     // this method is used to update a ticket assignee
     @Override
-    public Result addTicketAssignee(long ticketId, long assigneeId) {
+    public Result addTicketAssignee(long ticketId, long assigneeId, String token) {
+        if(!jwtTokenProvider.validateToken(token)){
+            return new Result(false, "Token is not valid");
+        }
+        if (!jwtTokenProvider.getRoleFromToken(token).equals("ADMIN")){
+            return new Result(false, "You are not authorized");
+        }
+
         Ticket newTicket = ticketDAO.findById(ticketId).get();
         newTicket.setAssignedTo(User.builder().id(assigneeId).build());
 
@@ -89,7 +114,14 @@ public class TicketManager implements TicketServices {
 
     // this method is used to remove a ticket assignee
     @Override
-    public Result removeTickerAssignee(long ticketId) {
+    public Result removeTickerAssignee(long ticketId, String token) {
+        if (!jwtTokenProvider.validateToken(token)){
+            return new Result(false, "Token is not valid");
+        }
+        if (!jwtTokenProvider.getRoleFromToken(token).equals("ADMIN")){
+            return new Result(false, "You are not authorized");
+        }
+
         Ticket newTicket = ticketDAO.findById(ticketId).get();
         newTicket.setAssignedTo(null);
         try {
